MorpheusService is used exclusively by the GreekService now.
